Split up spec by actions
